# perf(gatsby-plugin-mdx): convert jsx to js with sucrase, not babel