--- a/packages/gatsby-plugin-mdx/package.json
+++ b/packages/gatsby-plugin-mdx/package.json
@@ -49,6 +49,7 @@
     "slugify": "^1.4.4",
     "static-site-generator-webpack-plugin": "^3.4.2",
     "style-to-object": "^0.3.0",
+    "sucrase": "^3.16.0",
     "underscore.string": "^3.3.5",
     "unified": "^8.4.2",
     "unist-util-map": "^1.0.5",