--- a/packages/gatsby-plugin-mdx/utils/gen-mdx.js
+++ b/packages/gatsby-plugin-mdx/utils/gen-mdx.js
@@ -1,14 +1,10 @@
-const babel = require(`@babel/core`)
+const sucrase = require(`sucrase`)
 const grayMatter = require(`gray-matter`)
 const mdx = require(`@mdx-js/mdx`)
-const objRestSpread = require(`@babel/plugin-proposal-object-rest-spread`)
 
 const debug = require(`debug`)(`gatsby-plugin-mdx:gen-mdx`)
 
 const getSourcePluginsAsRemarkPlugins = require(`./get-source-plugins-as-remark-plugins`)
-const htmlAttrToJSXAttr = require(`./babel-plugin-html-attr-to-jsx-attr`)
-const removeExportKeywords = require(`./babel-plugin-remove-export-keywords`)
-const BabelPluginPluckImports = require(`./babel-plugin-pluck-imports`)
 const { parseImportBindings } = require(`./import-parser`)
 
 /*
@@ -128,6 +124,7 @@ export const _frontmatter = ${JSON.stringify(data)}`
   )
 
   debug(`running mdx`)
+  // TODO: if we can somehow drop the imports/exports in this call we can remove the regexes below
   let code = await mdx(content, {
     filepath: node.fileAbsolutePath,
     ...options,
@@ -142,40 +139,10 @@ ${code}`
 
   if (!isLoader) {
     debug(`compiling scope`)
-    const instance = new BabelPluginPluckImports()
-    const result = babel.transform(code, {
-      configFile: false,
-      plugins: [
-        instance.plugin,
-        objRestSpread,
-        htmlAttrToJSXAttr,
-        removeExportKeywords,
-      ],
-      presets: [
-        require(`@babel/preset-react`),
-        [
-          require(`@babel/preset-env`),
-          {
-            useBuiltIns: `entry`,
-            corejs: 3,
-            modules: false,
-          },
-        ],
-      ],
-    })
 
-    const identifiers = Array.from(instance.state.identifiers)
-    const imports = Array.from(instance.state.imports)
-    if (!identifiers.includes(`React`)) {
-      identifiers.push(`React`)
-      imports.push(`import * as React from 'react'`)
-    }
-
-    results.scopeImports = imports
-    results.scopeIdentifiers = identifiers
-    // TODO: be more sophisticated about these replacements
-    results.body = result.code
-      .replace(
+    results.body = sucrase
+      .transform(code, { transforms: [`jsx`] })
+      .code.replace(
         /export\s*default\s*function\s*MDXContent\s*/,
         `return function MDXContent`
       )
@@ -183,6 +150,9 @@ ${code}`
         /export\s*{\s*MDXContent\s+as\s+default\s*};?/,
         `return MDXContent;`
       )
+      .replace(/^(\s*import\s+.*)/gm, `/*** $1 ***/`)
+      .replace(/^(\s*export\s+\{.*)/gm, `/*** $1 ***/`)
+      .replace(/^(\s*export\s+)(.*)/gm, `/*** $1 ***/ $2`)
   }
   /* results.html = renderToStaticMarkup(
    *   React.createElement(MDXRenderer, null, results.body)