This is not a stable PR. Do not publish.

Currently the last super slow that is step left in the mdx build pipeline is the conversion of jsx source to js source code. The mdx call will return a jsx source code string but we need plain js because we `eval` execute the code in nodejs later, so we can't have the jsx.

In addition, the imports and exports are also eliminated from the code. Though that _could_ be done without a parser (and in fact, this PR does that).

Right now the only feasible way of transforming the jsx out is to actually parse the source code we receive from mdx again and to transform the jsx out. Currently this is done through Babel. But that's super slow.

This PR changes this step to convert the jsx through Sucrase.

Some numbers; in a 100k build the run queries step took 2973s. Of that, 2058s is attributed to the body resolver of mdx and 1279s of that was attributed to running `babel.transform`. With sucrase the run queries step is 1113s, the body resolver 538s, and sucrase is 24s.

Unfortunately the new approach will break any code that _depends_ on the way that Babel transforms the code. Additionally, it uses regular expressions to remove the `import` and `export` keywords, assuming a line can only start with them (and a space) if they are in fact ESM declarations. This could only be false with multi-line templates, multi-line comments, or invalid code. I don't think those are generated, but I still need to validate that assumption. Furthermore it may break other assumptions, so the switch is non-trivial and not free of risk.